+++ templates/service-express/server/src/services/pong.service.ts
@@ -1,8 +1,5 @@
 import type { PingService } from './ping.service';
 import { injectService, service } from 'nstarter-core';
-//#module apm
-import { apmTransaction } from 'nstarter-apm';
-//#endmodule apm
 
 @service()
 export class PongService {
@@ -16,11 +13,4 @@ export class PongService {
     public ping() {
         this.pingService.ping();
     }
-
-    //#module apm
-    @apmTransaction()
-    public apmTrans() {
-        console.log('apm');
-    }
-    //#endmodule apm
 }